pyproject.toml
[project]
name = "legal-ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "openai>=2.31.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
